Change barrier implementation to match broadcast
The previous version didn't work properly when the same barrier was
used twice in a row, because the second invocation's `fetch_add`s
would race with the first invocation's `fetch_sub`s, and the `start`
variable wouldn't be reset properly as a result.

This version still requires the barrier to be zeroed initially, but
uses half the storage and fewer atomic operations on shared memory.

The static EPOCH variable may need to be replaced if a single shared
library is reused by multiple processes, but we won't worry about
that for now.